Split file per registry API
